changes to av_velocity  small cahnge#

+++ d2q9-bgk.c
@@ -369,7 +369,7 @@ int collision(const t_param params, t_speed* cells, t_speed* tmp_cells, int* obs
 double av_velocity(const t_param params, t_speed* cells, int* obstacles)
 {
   int tot_cells = 0;  /* no. of cells used in calculation */
-  float tot_u = 0.0;          /* accumulated magnitudes of velocity for each cell */
+  double tot_u = 0.0;          /* accumulated magnitudes of velocity for each cell */
 
   /* initialise */
   /* loop over all non-blocked cells */
@@ -384,7 +384,7 @@ double av_velocity(const t_param params, t_speed* cells, int* obstacles)
       {
         int cellAccess = ii * params.nx + jj;
 
-        float local_density = cells[cellAccess].speeds[0]
+        double local_density = cells[cellAccess].speeds[0]
                               +cells[cellAccess].speeds[1]
                               +cells[cellAccess].speeds[2]
                               +cells[cellAccess].speeds[3]
@@ -396,7 +396,7 @@ double av_velocity(const t_param params, t_speed* cells, int* obstacles)
 
 
         /* x-component of velocity */
-        float u_x = (cells[cellAccess].speeds[1]
+        double u_x = (cells[cellAccess].speeds[1]
                       + cells[cellAccess].speeds[5]
                       + cells[cellAccess].speeds[8]
                       - (cells[cellAccess].speeds[3]
@@ -404,7 +404,7 @@ double av_velocity(const t_param params, t_speed* cells, int* obstacles)
                          + cells[cellAccess].speeds[7]))
                      / local_density;
         /* compute y velocity component */
-        float u_y = (cells[cellAccess].speeds[2]
+        double u_y = (cells[cellAccess].speeds[2]
                       + cells[cellAccess].speeds[5]
                       + cells[cellAccess].speeds[6]
                       - (cells[cellAccess].speeds[4]
